requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
supabase>=1.0.3
python-dateutil>=2.8.2
aiohttp>=3.8.0
PyNaCl>=1.5.0
pytz>=2023.3 